requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alpaca-py==0.30.1
alpaca-trade-api==3.2.0
annotated-types==0.7.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
deprecation==2.1.0
fonttools==4.53.1
frozendict==2.4.4
frozenlist==1.4.1
html5lib==1.1
idna==3.8
importlib_resources==6.4.4
kiwisolver==1.4.5
lxml==5.3.0
matplotlib==3.9.2
msgpack==1.0.3
multidict==6.0.5
multitasking==0.0.11
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pandas_ta==0.3.14b0
peewee==3.17.6
pillow==10.4.0
platformdirs==4.2.2
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
soupsieve==2.6
sseclient-py==1.8.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
yarl==1.9.4
yfinance==0.2.43
zipp==3.20.1
